context code restructuring and other edits
re-organise context code for understandability

some refinements along the way, such as a new `context->toScreenCoordinates` API
